moved threads into its own file, removed unused imports, renamed lengthy variables in midi_detect
