+++ src/form.py
@@ -8,40 +8,22 @@ Wave Generation
 After defining the constants, the code generates waveforms for each key using different oscillators 
 (sine, square, sawtooth, and triangle). These waveforms are stored in dictionaries for easy access.
 
-Worker Class
-The Worker class is a helper class that is used to run the audio processing in a separate thread 
-to avoid blocking the GUI.
 
 MainWidget Class
 The MainWidget class represents the main widget of the synthesizer application. It inherits from 
 the QWidget class provided by the PySide6 library. The class contains methods for handling UI events, 
 such as button presses, knob changes, and waveform selection.
 
-MidiInputWorker Class
-The MidiInputWorker class is a part of a synthesizer GUI application written in Python using the 
-PySide6 library. 
-This class is responsible for handling MIDI input messages and connecting them to the generation of 
-tones in the synthesizer.
 
-MidiThread Class:
-The MidiThread class is responsible for managing the MIDI input functionality in a separate thread. 
-It initializes the MIDI system using pygame.midi.init() and defines a signal named start_midi_thread. 
-This class is designed to work with the MidiInputWorker class to receive and process MIDI messages
-concurrently without blocking the main user interface.
 """
 
 import os
-import sys
 from pathlib import Path
 from PySide6.QtWidgets import (
     QWidget,
-    QFrame,
     QPushButton,
-    QRadioButton,
-    QMessageBox,
-    QApplication,
 )
-from PySide6.QtCore import QFile, Qt, QObject, QRunnable, Slot, QThreadPool, Signal
+from PySide6.QtCore import QFile, QThreadPool
 from PySide6.QtUiTools import QUiLoader
 from oscillator import (
     SineOscillator as sine,
@@ -52,8 +34,8 @@ from oscillator import (
 from adsr import ADSREnvelope, State
 from notefreq import NOTE_FREQS
 from volume import Volume
-from midi_detect import identify_and_select_midi_device
-from midi_detect import receive_midi_input
+from midi_detect import identify_device
+from threads import Worker, MidiInputWorker, MidiWorker
 import pygame
 import sounddevice as sd
 import numpy as np
@@ -108,60 +90,6 @@ GUI_KEY_NAMES = [
 ]
 
 
-# Thread worker
-class Worker(QRunnable):
-    def __init__(self, fn, *args, **kwargs):
-        super(Worker, self).__init__()
-        # Store constructor arguments (re-used for processing)
-        self.fn = fn
-        self.args = args
-        self.kwargs = kwargs
-
-    @Slot()  # QtCore.Slot
-    def run(self):
-        self.fn(self.args[0])
-
-
-# Definition of MidiThread class that inherits QObject
-class MidiThread(QObject):
-    pygame.midi.init()
-    # Define the start_midi_thread signal
-    start_midi_thread = Signal()  # establishes a signal to manipulate
-
-    def __init__(
-        self, input_device
-    ):  # construct that is used for instances of MIDITHREAD class
-        super().__init__()
-        pygame.midi.init()
-        self.input_device = input_device
-
-    def start(self):
-        self.start_midi_thread.emit()
-
-class MidiInputWorker(QRunnable):
-    def __init__(self, input_device, main_widget):
-        super(MidiInputWorker, self).__init__()
-        
-        self.input_device = input_device
-        self.main_widget = main_widget
-
-    @Slot()
-    def run(self):
-        for midi_message in receive_midi_input(self.input_device):
-            # print("Yo! MIDI message:", midi_message) # debag line
-            
-            if midi_message["status"] == 144: # This is the note on message
-                note_value = midi_message["note"]
-                if note_value >= 12 and note_value < 122: 
-                    try:
-                        note_name = self.main_widget.pitch_shifted_keys[note_value - 24]
-                        print("MIDI KEY NOTE PLAYED:", note_name)
-                        self.main_widget.button_pressed_handler(note_name)
-                    except IndexError:
-                        print("Note value is out of range. Ignoring MIDI message.")
-                    
-            elif midi_message["status"] == 128: # note off message
-                self.main_widget.button_released_handler() 
 
 
 class MainWidget(
@@ -187,7 +115,7 @@ class MainWidget(
         # MIDI stuff here begins here:
         pygame.midi.init()
         input_device = (
-            identify_and_select_midi_device()
+            identify_device()
         )  # call device detection function once, and store it in Input_device variable
 
         #handling blurb for no-device situation 
@@ -201,7 +129,7 @@ class MainWidget(
             # Start the MidiInputWorker as a new thread
             self.threadpool.start(self.midi_worker)
 
-            self.midi_thread = MidiThread(None)
+            self.midi_thread = MidiWorker(None)
             # self.midi_thread.start_midi_thread.connect(lambda: self.midi_thread.receive_midi_input(input_device))
 
             # Start the MIDI thread
